+++ src/mimic/src/interface/mod.rs
@@ -1,14 +1,12 @@
 pub mod query;
 
-use candid::CandidType;
-use serde::{Deserialize, Serialize};
 use thiserror::Error as ThisError;
 
 ///
 /// InterfaceError
 ///
 
-#[derive(CandidType, Debug, Serialize, Deserialize, ThisError)]
+#[derive(Debug, ThisError)]
 pub enum InterfaceError {
     #[error(transparent)]
     QueryError(#[from] query::QueryError),