+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.4.3]
 - moved base out of mimic and added mimic_base.  mimic_common can now go back into mimic
 - Sorted moved to a trait, which altered the logic so that traits can have attributes but no impl or derive
+- moved build back into mimic
+- changed the Errors so only the top level MimicError derives CandidType/Serialize/Deserialize
 
 ## [0.4.2]
 - made Db into StoreRegistry and made it generic.  Now you can define multiple registries (one for DataStores, one